Moves Excel download components to separate file
